Draft GearProofStorage
